Preparación proeycto para manejar bases de datos con mongoose, manejo de módulos y organización de la estructura general del proyecto
