+++ package.json
@@ -3,11 +3,16 @@
   "version": "1.0.0",
   "main": "src/app",
   "dependencies": {
+    "bcrypt": "^3.0.6",
     "body-parse": "^0.1.0",
     "bootstrap": "^4.3.1",
     "express": "^4.17.0",
+    "express-session": "^1.16.2",
     "hbs": "^4.0.4",
     "jquery": "^3.4.1",
+    "memorystore": "^1.6.1",
+    "mongoose": "^5.6.2",
+    "mongoose-unique-validator": "^2.0.3",
     "popper.js": "^1.15.0",
     "yargs": "^13.2.4"
   },
