+++ src/estudiantes.json
@@ -1,1 +1,1 @@
-[{"documento":0,"nombre":"Admin","correo":"admin@nodejs.com","telefono":"123","contrasena":"1","tipo":"coordinador"},{"documento":123,"correo":"a@q","nombre":"Walter2222666","telefono":"123666","contrasena":"123","tipo":"aspirante"},{"documento":456,"correo":"a@q9","nombre":"Jhon933","telefono":"4566669","contrasena":"123","tipo":"aspirante"},{"documento":789,"correo":"a@q","nombre":"Sebas1","telefono":"789666","contrasena":"123","tipo":"docente"},{"documento":15515515,"nombre":"Walter Serna Zapata1","correo":"walterasz4@gmail.com","telefono":"3002745528","contrasena":"123","tipo":"docente"},{"documento":71223998,"nombre":"Jhon Edison","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"aspirante"},{"documento":71223,"nombre":"Jhon Edison819","correo":"jems10@hotmail.com819","telefono":"3215516506819","contrasena":"123","tipo":"docente"},{"documento":90909090,"nombre":"P","correo":"asdawd@asdawd.awaw","telefono":"3","tipo":"aspirante"},{"documento":1221,"nombre":"LLLLLL","correo":"nam@as.we","telefono":"90999090","contrasena":"1221","tipo":"aspirante"},{"documento":5555,"nombre":"2222","correo":"asdawdawdl@asdaw.aw","telefono":"2222","tipo":"aspirante"},{"documento":45555,"nombre":"12233","correo":"asdalwhdaw@asdaw.aw","telefono":"2","tipo":"aspirante"},{"documento":2332,"nombre":"E","correo":"A@a.a","telefono":"1","contrasena":"1","tipo":"aspirante"},{"documento":121212,"nombre":"3","correo":"a@a2.1","telefono":"d","contrasena":"1","tipo":"aspirante"},{"documento":123333,"nombre":"Sergio","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"docente"},{"documento":8888,"nombre":"Segundo","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"docente"},{"documento":989,"nombre":"PruebaFinal","correo":"asdawda@asdawa.aw","telefono":"9","contrasena":"989","tipo":"docente"},{"documento":1234,"nombre":"OTRA PRUEBA EDICION","correo":"abcEDICION@abc.a","telefono":"3000","contrasena":"1234","tipo":"aspirante"}]
+[{"documento":0,"nombre":"Admin","correo":"admin@nodejs.com","telefono":"123","contrasena":"1","tipo":"coordinador"},{"documento":123,"correo":"a@q","nombre":"Walter2222666","telefono":"123666","contrasena":"123","tipo":"aspirante"},{"documento":456,"correo":"a@q9","nombre":"Jhon933","telefono":"4566669","contrasena":"123","tipo":"aspirante"},{"documento":789,"correo":"a@q","nombre":"Sebas1","telefono":"789666","contrasena":"123","tipo":"docente"},{"documento":15515515,"nombre":"Walter Serna Zapata1","correo":"walterasz4@gmail.com","telefono":"3002745528","contrasena":"123","tipo":"docente"},{"documento":71223998,"nombre":"Jhon Edison","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"aspirante"},{"documento":71223,"nombre":"Jhon Edison819","correo":"jems10@hotmail.com819","telefono":"3215516506819","contrasena":"123","tipo":"docente"},{"documento":90909090,"nombre":"P","correo":"asdawd@asdawd.awaw","telefono":"3","tipo":"aspirante"},{"documento":1221,"nombre":"LLLLLL","correo":"nam@as.we","telefono":"90999090","contrasena":"1221","tipo":"aspirante"},{"documento":5555,"nombre":"2222","correo":"asdawdawdl@asdaw.aw","telefono":"2222","tipo":"aspirante"},{"documento":45555,"nombre":"12233","correo":"asdalwhdaw@asdaw.aw","telefono":"2","tipo":"aspirante"},{"documento":2332,"nombre":"E","correo":"A@a.a","telefono":"1","contrasena":"1","tipo":"aspirante"},{"documento":121212,"nombre":"3","correo":"a@a2.1","telefono":"d","contrasena":"1","tipo":"aspirante"},{"documento":123333,"nombre":"Sergio","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"docente"},{"documento":8888,"nombre":"Segundo","correo":"jems10@hotmail.com","telefono":"3215516506","contrasena":"123","tipo":"docente"},{"documento":989,"nombre":"PruebaFinal","correo":"asdawda@asdawa.aw","telefono":"9","contrasena":"989","tipo":"docente"},{"documento":1234,"nombre":"OTRA PRUEBA EDICION","correo":"abcEDICION@abc.a","telefono":"3000","contrasena":"1234","tipo":"aspirante"},{"documento":9090,"nombre":"9090","correo":"9090@asdawd.aw","telefono":"9090","contrasena":"9090","tipo":"aspirante"}]
