--- src/listado - copia.json
@@ -1,1 +1,0 @@
-[{"nombre":"Mariana","matematicas":4,"ingles":1,"programacion":2},{"nombre":"ana","matematicas":2,"ingles":1,"programacion":2},{"nombre":"juan","matematicas":5,"ingles":3,"programacion":2}]
