+++ template/views/eliminar-curso-confirmacion.hbs
