+++ template/views/eliminar-curso-docente-confirmacion.hbs
