+++ template/views/inscripcion-confirmacion.hbs
