+++ template/views/registrar-usuario.hbs
